+++ src/view/components/FormInputDate.js
@@ -5,18 +5,15 @@ export default function FormInputText({
   tip,
   content,
   setContent,
-  placeholder,
   propertyID,
 }) {
   return (
     <div className="form-group">
       <label htmlFor={propertyID}>{title}</label>
       <input
-        type="text"
+        type="date"
         className="form-control"
         id={propertyID}
-        aria-describedby={`${propertyID}Help`}
-        placeholder={placeholder}
         value={content}
         onChange={(e) => setContent(e.target.value)}
       />