+++ src/view/components/ResetPasswordForm.js
@@ -1,23 +1,22 @@
 import React from "react";
 
-import FormInputText from "./FormInputText";
+import FormInputPatternFullName from "./FormInputPatternFullName";
 import FormInputSwitch from "./FormInputSwitch";
 import FormInputFinalButton from "./FormInputFinalButton";
 
 export default function ResetPasswordForm({
   handlePasswordReset,
-  username,
-  setUsername,
+  fullName,
+  setFullName,
   setHasAcceptedTermsOfUse,
 }) {
   return (
-    <form id="ResetPasswordForm" className="col-10 col-md-6" onSubmit={handlePasswordReset}>
-      <FormInputText
-        username={username}
-        setUsername={setUsername}
-        labelContent="Seu nome de identificação no sistema"
-        propertyID="userNameForNewPassword"
-      />
+    <form
+      id="ResetPasswordForm"
+      className="col-10 col-md-6"
+      onSubmit={handlePasswordReset}
+    >
+      <FormInputPatternFullName content={fullName} setContent={setFullName} />
       <FormInputSwitch
         setHasAcceptedTermsOfUse={setHasAcceptedTermsOfUse}
         labelContent="Estou ciente que a senha antiga será substituída"